Backend/requirements.txt
# Data Processing & Parsing
pandas==2.1.3
openpyxl==3.1.2
lxml==4.9.3
beautifulsoup4==4.12.2

# HTTP & Networking
requests==2.31.0
urllib3==1.26.18
fake-useragent==1.4.0

# API & Web Frameworks
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
flask==2.3.3
flask-cors==4.0.0
blinker==1.6.2

# Browser Automation (Selenium) - STEALTH MODE
selenium==4.11.2
webdriver-manager==4.0.1
selenium-stealth==1.0.6
selenium-wire==5.1.0
undetected-chromedriver>=3.5.0  # Cloudflare, Turnstile bypass
cloudscraper

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
psycopg2-binary>=2.9.0  # PostgreSQL driver

# Security & Utilities
python-dotenv==1.0.0
cryptography==41.0.7
certifi==2023.11.17
packaging==23.2
setuptools==68.2.2

# Authentication
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
email-validator>=2.0.0